In a huge breach of established convention, renamed VELNSView to NSVelvetView.
Although this is technically invading Apple's namespace, they're unlikely to use the name "Velvet" for anything, and I believe that drawing a very clear distinction between AppKit responsibilities and Velvet responsibilities will be useful. In other words, anything starting with "NS" is understood to be managed by AppKit, while anything starting with "VEL" is understood to be managed by the Velvet framework.